refactor: move location of libplanet library
