+++ libplanet/merkle_trie_node.go
