+++ libplanet/merkle_trie_proof.go
