+++ libplanet/merkle_trie_proof_utils.go
